TandemBicycle.py
So um i usually say "this is my code" so i dont want to say that this time. I'm gonna say today:

So um i usually say "this is my code" so i dont want to say that this time. I'm gonna say today:
This is my code. Thank you.

Hehehehehehe :)