Object_sample.py
This is not related to my ccc class but i decided to put it here because in python I was learning classes and stuff so I wanted to show whoever out there is reading my code (probably no one cuz this ccc class folder thingy is private lol but we no talk about that hehehe). This is a code that I wrote as an example of the uses of classes. I hope you enjoy reading it and ya that's about all I want to say (quite a lot I know sorry but just wanted you to know i should stop writing now or else who is reading this won't want to read it anymore so byeeeeeeeeeeeeeee)