Cheesecake_Distribution.py and SteamArithmetic.py
These are my two codes. The cheesecake one is right but the steam one has something wrong, I can't figure it out tho, I'M SO CONFUSSSED. And btw for the discription ZeroThatOut.py and Crayola.py, I said BaNanA Pee; but that was a mistake it was supposed to be peel okay sorry 'bout that. Thank you.