ModeFinding.py and UniqueElements.py
2020/8/17
So today I added the date! I decided it would be harder to find the date than look at the descrivtion and know (expecially since it doesn't even show the year), so I put it in here. And also, I DID 2 WHOLE CCC QUESTIONSSS. Okay, I get it. Less impressive than 5 (like late time), but STILL. That's all for this lesson, thank you for watching educator.com. Lol that's usually the ending for a lesson on educator.com. But actually it's then end of this deScRipTiOn. So byeeee