PhoneBook.py
So this is phonebook.py. Just a note, for the rest couple of days, i doubt i will have anything to submit because i did all the ccc questions for this week already, and i submitted it all too. So... that's why i probably won't submit many things. Okay, tha's it. Thank you.