Create Frame.py
So... this was one of the questions in our quiz today. Actually most of the questions weren't as hard as I thought ngl but hey, there's nothing to complain about that is there? Anyways... THANK YOU FOR STAYING FOR MY TREATS ALBUS!!! Ok jk that was just what Scorpuis said to Albus in the Cursed Child. Anyways anyways, that's all. Thank you :)