Summary (required)
Description. Hehe. So I technically did 3 dmoooooooji problems today, since 2 of them I did in the past, and one I did today. This week, I won't be making a notebook, because it takes A LOT of time just to open, and sometimes the output is not what it would have been. And thirdy, it's really hard to read my code on GitHub in a notebook so yeah. :))))))  thanks. Bye by buy! (lol)