More Moving
So i moved all the code i did last ccc term into "lvl 1" so that's why all these stuff are here. Thank you.
P.S. CCC contest is tomorrow! as in tomorrow tomorrow! as in tomorrow tomorrow tomorrow! as in tomorrow tomorrow tomorrow tomorrow! as in—you get the idea.